requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.50.0
django-crispy-forms==1.13.0
django-embed-video==1.4.1
django-social-share==2.2.1
django-taggit==2.1.0
django-tinymce==3.4.0
django-tinymce4-lite==1.8.0
idna==3.3
jsmin==3.0.1
oauthlib==3.2.0
Pillow==9.0.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.8
